+++ ui/packages/shared/profile/src/MatchersInput/index.tsx
@@ -46,7 +46,7 @@ export interface ILabelNamesResult {
 interface UseLabelNames {
   result: ILabelNamesResult;
   loading: boolean;
-  refetch: () => void;
+  refetch: () => Promise<void>;
 }
 
 export const useLabelNames = (
@@ -85,8 +85,8 @@ export const useLabelNames = (
   return {
     result: {response: data, error: error as Error},
     loading: isLoading,
-    refetch: () => {
-      void refetch();
+    refetch: async () => {
+      await refetch();
     },
   };
 };
@@ -97,7 +97,7 @@ interface UseLabelValues {
     error?: Error;
   };
   loading: boolean;
-  refetch: () => void;
+  refetch: () => Promise<void>;
 }
 
 export const useLabelValues = (
@@ -135,8 +135,8 @@ export const useLabelValues = (
   return {
     result: {response: data ?? [], error: error as Error},
     loading: isLoading,
-    refetch: () => {
-      void refetch();
+    refetch: async () => {
+      await refetch();
     },
   };
 };