+++ ui/packages/app/web/CHANGELOG.md
@@ -3,6 +3,34 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.16.954](https://github.com/parca-dev/parca/compare/@parca/web@0.16.953...@parca/web@0.16.954) (2025-11-03)
+
+**Note:** Version bump only for package @parca/web
+
+## [0.16.953](https://github.com/parca-dev/parca/compare/@parca/web@0.16.952...@parca/web@0.16.953) (2025-10-28)
+
+**Note:** Version bump only for package @parca/web
+
+## [0.16.952](https://github.com/parca-dev/parca/compare/@parca/web@0.16.951...@parca/web@0.16.952) (2025-10-28)
+
+**Note:** Version bump only for package @parca/web
+
+## [0.16.951](https://github.com/parca-dev/parca/compare/@parca/web@0.16.950...@parca/web@0.16.951) (2025-10-22)
+
+**Note:** Version bump only for package @parca/web
+
+## [0.16.950](https://github.com/parca-dev/parca/compare/@parca/web@0.16.949...@parca/web@0.16.950) (2025-10-21)
+
+**Note:** Version bump only for package @parca/web
+
+## [0.16.949](https://github.com/parca-dev/parca/compare/@parca/web@0.16.948...@parca/web@0.16.949) (2025-10-21)
+
+**Note:** Version bump only for package @parca/web
+
+## [0.16.948](https://github.com/parca-dev/parca/compare/@parca/web@0.16.947...@parca/web@0.16.948) (2025-10-21)
+
+**Note:** Version bump only for package @parca/web
+
 ## [0.16.947](https://github.com/parca-dev/parca/compare/@parca/web@0.16.946...@parca/web@0.16.947) (2025-10-16)
 
 **Note:** Version bump only for package @parca/web