+++ ui/packages/shared/components/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 0.16.380 (2025-11-03)
+
+**Note:** Version bump only for package @parca/components
+
+## 0.16.379 (2025-10-22)
+
+**Note:** Version bump only for package @parca/components
+
+## [0.16.378](https://github.com/parca-dev/parca/compare/@parca/components@0.16.377...@parca/components@0.16.378) (2025-10-21)
+
+**Note:** Version bump only for package @parca/components
+
+## 0.16.377 (2025-10-21)
+
+**Note:** Version bump only for package @parca/components
+
 ## [0.16.376](https://github.com/parca-dev/parca/compare/@parca/components@0.16.375...@parca/components@0.16.376) (2025-10-08)
 
 **Note:** Version bump only for package @parca/components