+++ ui/packages/shared/utilities/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 0.0.113 (2025-11-03)
+
+**Note:** Version bump only for package @parca/utilities
+
+## 0.0.112 (2025-10-21)
+
+**Note:** Version bump only for package @parca/utilities
+
 ## 0.0.111 (2025-10-02)
 
 **Note:** Version bump only for package @parca/utilities